Moved post tests from shell script to pytest
