--- tests/test_feed/.gitignore
@@ -1,3 +1,0 @@
-*.db
-*.log
-*.out
