--- tests/test_post/.gitignore
@@ -1,3 +1,0 @@
-*.db
-*.log
-*.out
